% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{compile_data}
\alias{compile_data}
\title{Compile data and meta-data to wide format}
\usage{
compile_data(dat, meta, na.rm = TRUE)
}
\arguments{
\item{dat}{A MarineDatabase data object. A data file that has passed through one of the process data functions in the MarineDatabase package, such as \code{\link{process_cdom_data}}.}

\item{meta}{A \code{\link[=export_metadata]{MetaData}} object.}

\item{na.rm}{Logical indicating whether sample names that are missing from meta-data should be removed. Defaults TRUE.}
}
\value{
Returns a data frame that contains all entries in \code{dat} that could be \code{\link{merge}}d with \code{meta}.
}
\description{
Compiles data with meta-data using the standardized sample names. The compiled datasets can be readily exported as Excel sheets.
}
\details{
The \code{sample_name} column is used to merge data. Sample names in \code{dat} and \code{meta} have to match and be unique.
}
\author{
Mikko Vihtakari
}
