% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_abundance.R
\name{convert_abundance}
\alias{convert_abundance}
\title{Calculate and convert between abundance values for ice core and planton net data}
\usage{
convert_abundance(data, ab_col, ab_from = NULL, ab_to = NULL,
  filtered = NULL, type = "icecore", cf = 1, vol_col = NULL,
  start_col = "from", end_col = "to", unit_col = "unit")
}
\arguments{
\item{data}{Data.frame containing the data to be converted}

\item{ab_col}{Character vector specifying the columns containing abundance data}

\item{ab_from}{The initial abundance unit. Options: "n" (number of individuals), "nsub" (number of individuals in a subsample), "1/L" (number of individuals per litre), "1/m3" (number of individuals per cubic metre), "1/m2" (number of individuals per square metre).}

\item{ab_to}{The desired abundance unit. Options: "1/L" (number of individuals per litre), "1/m3" (number of individuals per cubic metre), "1/m2" (number of individuals per square metre).}

\item{filtered}{Logical indicating whether the sample was filtered (\code{filtered = TRUE}) or subsampled (\code{filtered = FALSE}). See details.}

\item{type}{Character argument specifying either the gear column (see \code{\link{read_metadata}}) or either of following: "icecore" or "net".}

\item{cf}{Numeric giving the concentration factor connected to the volume difference when transforming water to ice (melted water volume/ice volume). Used only, if \code{type} indicates ice core. To remove the effect, set \code{cf = 1}.}

\item{vol_col}{Character argument specifying the volume column to be used in calculations. Should be supplied in \strong{liters}.}

\item{start_col}{Character argument specifying the name of the column in \code{data} for the section start for the ice core or plankton net in \strong{meters}.}

\item{end_col}{Character argument specifying the name of the column in \code{data} for the section end for the ice core or plankton net in \strong{meters}.}

\item{unit_col}{Character argument specifying the unit column name in \code{data}.}
}
\description{
Not ready conversion function
}
\details{
The function to calculate ind m^-2 values from ind is (ind*cf*1000*h)/V, where ind is number of individuals in a sample, cf the concentration factor, h height of ice-core section in m, and V melted volume in L.

The function is not ready and works only for "1/L" to "1/m2" conversions with very limited options. 

\itemize{
  \item \strong{n} 
  }
}
\author{
Mikko Vihtakari
}
