% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sampling_depths.R
\name{convert_sampling_depths}
\alias{convert_sampling_depths}
\title{Convert sampling depths to numeric.}
\usage{
convert_sampling_depths(dt, column, comment.col = "comment")
}
\arguments{
\item{dt}{data frame containing meta-data. See \code{\link{export_metadata}}}

\item{column}{column name in meta-data containing the sampling depths to be converted.}

\item{comment.col}{the name of the comment column in \code{dt}}
}
\description{
Converts sampling depths to numeric and adds any character strings from these columns to comments
}
\author{
Mikko Vihtakari
}
