% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ctd_kongsfjord}
\alias{ctd_kongsfjord}
\title{Example CTD data for Kongsfjorden}
\format{
data.frame
}
\usage{
data("ctd_kongsfjord")
}
\description{
A dataframe containing example CTD data for Kongsfjorden.
}
