% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ctd_rijpfjord}
\alias{ctd_rijpfjord}
\title{Example CTD data for Rijpfjorden}
\format{data.frame}
\usage{
data("ctd_rijpfjord")
}
\description{
A dataframe containing example CTD data for Rijpfjorden.
}
