% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_water_type.R
\name{define_water_type}
\alias{define_water_type}
\title{Define water type from temparature and salinity for Kongsfjorden}
\usage{
define_water_type(dt, temp_col = "temp", sal_col = "sal", bind = FALSE,
  return_water_type_list = FALSE)
}
\arguments{
\item{dt}{Dataframe containing temperature and salinity information}

\item{temp_col}{Character string giving the column name containing temperature data}

\item{sal_col}{Character string giving the column name containing salinity data}

\item{bind}{Should the water type information be \code{\link{cbind}} with \code{dt} (\code{TRUE}) or should a character vector containing water type abbreviations returned instead (\code{FALSE}; default)}

\item{return_water_type_list}{Should dataframe containing water type definitions be returned instead of defining water types (\code{TRUE}; default is \code{FALSE})?}
}
\description{
Defines water type from temperature and salinity data for Kongsfjorden
}
\details{
Water types are returned after the definition in Cottier et al. (2005) with slight modifications to avoid overlaps of water types. Use \code{return_water_type_list = TRUE} to see the boundary values for water masses. \code{y} means temperature and \code{x} salinity. These values are used in associated TS diagrams.
}
\references{
Cottier, F., Tverberg, V., Inall, M., Svendsen, H., Nilsen, F., Griffiths, C., 2005. Water mass modification in an Arctic fjord through cross-shelf exchange: The seasonal hydrography of Kongsfjorden, Svalbard. J. Geophys. Res. 110, C12005. doi:10.1029/2004JC002757
}
\author{
Mikko Vihtakari
}
