% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_station_names.R
\name{fix_station_names}
\alias{fix_station_names}
\title{Match station names against standardized station names for Kongsfjorden and Ripfjorden}
\usage{
fix_station_names(x)
}
\arguments{
\item{x}{A character vector or factor to be matched}
}
\value{
Returns a names character vector of corrected station names
}
\description{
Matches and fixes inconsistencies in station names for NPI's standard marine biological stations
}
\author{
Mikko Vihtakari
}
