% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ext.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Get file extension}
\source{
The function is borrowed from StackOverflow (\url{https://stackoverflow.com/a/45467246/1082004})
}
\usage{
get_file_ext(fn)
}
\arguments{
\item{fn}{File name as a character vector}
}
\description{
Extracts file extension from a file name
}
\author{
Pisca46
}
