% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spabbr.R
\name{make_spabbr}
\alias{make_spabbr}
\title{Make species abbreviations from latin species names}
\usage{
make_spabbr(x)
}
\arguments{
\item{x}{character vector containing latin species names to be abbreviated. Taxonomic levels must be separated by whitespace.}
}
\value{
Returns a character vector containing species abbreviations. The abbreviations should be valid data.frame column names. Check the uniquety of the returned names (that is not checked by the function).
}
\description{
The function generates species names from latin names
}
