% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_species_list.R
\name{make_species_list}
\alias{make_species_list}
\title{Generate a species list containing unique levels from a taxonomy dataset}
\usage{
make_species_list(dat, sp.col = "species", ab.col = "abundance",
  st.col = "station", expedition = NA)
}
\arguments{
\item{dat}{data frame containing taxonomy information. Must be in long format.}

\item{sp.col}{Character specifying the column name that contains species names in \code{dat}.}

\item{ab.col}{Character specifying the column name that contains species abundances in \code{dat}.}

\item{st.col}{Character specifying the column name that contains station names in \code{dat}.}

\item{expedition}{Character specifying the column name that contains expedition in \code{dat}. Not required. Defaults to \code{NA}.}
}
\value{
Returns a list containing a data frame (\code{$species_list}) the original species names (as unique levels in \code{dat}; or_name), summed abundance (ab), number of stations where the species name was encountered (n), total number of station (N), a tidied search term for XXX function (search_term), type of species (spores and epiphytes are separated, otherwise \code{NA}; type) and certainty of species identification (includes sp., spp., aff. and cf.; certainty)
}
\description{
Generate a data frame containing unique species names, summed abundances and search terms to be used in XXX
}
\author{
Mikko Vihtakari
}
