% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_station_data.R
\name{make_station_data}
\alias{make_station_data}
\title{Compile marine biological data to Norwegian Polar Institute data format}
\usage{
make_station_data(dt, station.col = "Station", date.col = "Date",
  expedition = "TWICE", latitude.col = "lat", longitude.col = "lon",
  ctds = NULL, variable.col = "variable", value.col = "value",
  type.col = "Type", gear.col = "Gear", unit.col = "unit",
  sample.name.col = "Name", sampling.depth.col = "From", find.ctds = TRUE)
}
\arguments{
\item{dt}{data frame containing overview of the cruise data. See details for formatting instructions.}

\item{station.col}{Character. Column name specifying the station name in \code{dt}}

\item{date.col}{Character. Column name specifying the sampling time in \code{dt}. The \code{date.col} in \code{dt} be in \code{\link[=DateTimeClasses]{POSIXct}} format.}

\item{expedition}{Character string specifying the name of the expedition.}

\item{latitude.col}{Character. Column name specifying the latitude column in \code{dt}. The \code{latitude.col} in \code{dt} be given in decimal degrees.}

\item{longitude.col}{Character. Column name specifying the longitude column in \code{dt}. The \code{longitude.col} in \code{dt} be given in decimal degrees.}

\item{ctds}{list of \link[oce]{ctd-class} objects taken at the station. Required if \code{dt} contains CTD casts.}

\item{variable.col}{Character. Column name specifying the variable column. This is stupid programming and should be changed.}

\item{value.col}{Character. Column name specifying the value column. This is stupid programming and should be changed.}

\item{type.col}{Column name specifying the sample type column in \code{dt}. This will be used to place data on the same level than CTD data in the hierarchy. Important!}

\item{gear.col}{Column name specifying the sampling gear column in \code{dt}}

\item{unit.col}{Column name specifying the unit of \code{value.col} column in \code{dt}}

\item{sample.name.col}{Column name specifying the sample name column in \code{dt}. Important for binding meta-data and actual data together.}

\item{sampling.depth.col}{Column name specifying the sampling depth column in \code{dt}. Important parameter. Should be given in meters.}

\item{find.ctds}{Logical. Should the function find ctd casts from a list of \link[oce]{ctd-class} objects based on station name?}
}
\description{
The function compiles scientific cruise data from +emph{one station} to the format required by the Norwegian Polar Institute data base
}
\details{
The data frame for \code{dt} should be formatted as Dr. Anette Wold's Excel sheets tend to be formatted.
}
\author{
Mikko Vihtakari
}
