% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_zooplankton_data.R
\name{merge_zooplankton_data}
\alias{merge_zooplankton_data}
\title{Merge two ZooplanktonData objects}
\usage{
merge_zooplankton_data(x, y)
}
\arguments{
\item{x}{\link[=read_zooplankton_data]{ZooplanktonData} object}

\item{y}{\link[=read_zooplankton_data]{ZooplanktonData} object}
}
\description{
Merges two \link[=read_zooplankton_data]{ZooplanktonData} objects
}
\seealso{
\code{\link{read_zooplankton_data}}

Other ZooplanktonData: 
\code{\link{print.ZooplanktonData}()},
\code{\link{read_zooplankton_data}()},
\code{\link{subset.ZooplanktonData}()},
\code{\link{summarize_zooplankton_data}()}
}
\author{
Mikko Vihtakari
}
\concept{ZooplanktonData}
