% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CDOMdata.R
\name{plot.CDOMdata}
\alias{plot.CDOMdata}
\title{Plot CDOM data}
\usage{
\method{plot}{CDOMdata}(x, type = "spectra")
}
\arguments{
\item{x}{\code{CDOMdata} object from \code{\link{process_cdom_data}} function.}

\item{type}{Character specifying the type of the plot. Options: \code{"spectra"}, \code{"models"} or \code{"model_fit"}. See Details}
}
\description{
Plot method for \code{\link[=process_cdom_data]{CDOMdata}} objects.
}
\details{
Three plot types have been implemented:
\itemize{
\item \code{\strong{"spectra"}} Plots the absorbtion coefficient of all samples against wavelength. Flags samples that had negative values in the range between 350 and 400 nm. 
\item \code{\strong{"models"}} Plots exponential decay models fitted using the \code{\link{nls}} function. See Value in \code{\link{process_cdom_data}} for details.
\item \code{\strong{"model_fit"}} Plots both the absorbition coefficients and fitted nonlinear models. 
}

Note that especially the \code{"model_fit"} option might be slow for large datasets.
}
\seealso{
\code{\link{process_cdom_data}}
}
\author{
Mikko Vihtakari
}
