% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CDOMdata.R
\name{print.CDOMdata}
\alias{print.CDOMdata}
\title{Print \code{CDOMdata} objects}
\usage{
\method{print}{CDOMdata}(x, ...)
}
\arguments{
\item{x}{\code{CDOMdata} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=process_cdom_data]{CDOMdata}} objects
}
\seealso{
\code{\link{process_cdom_data}}
}
\author{
Mikko Vihtakari
}
