% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MetaData.R
\name{print.MetaData}
\alias{print.MetaData}
\title{Print \code{MetaData} objects}
\usage{
\method{print}{MetaData}(x, ...)
}
\arguments{
\item{x}{\code{MetaData} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=export_metadata]{MetaData}} objects
}
\seealso{
\code{\link{export_metadata}}
}
\author{
Mikko Vihtakari
}
