% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NPIMarData.R
\name{print.NPIMarData}
\alias{print.NPIMarData}
\title{Print \code{NPIMarData} objects}
\usage{
\method{print}{NPIMarData}(x, ...)
}
\arguments{
\item{x}{\code{NPIMarData} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=make_station_data]{NPIMarData}} objects
}
\seealso{
\code{\link{make_station_data}}
}
\author{
Mikko Vihtakari
}
