% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ZooplanktonData.R
\name{print.ZooplanktonData}
\alias{print.ZooplanktonData}
\title{Print \code{ZooplanktonData} objects}
\usage{
\method{print}{ZooplanktonData}(x, ...)
}
\arguments{
\item{x}{\code{ZooplanktonData} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=read_zooplankton_data]{ZooplanktonData}} objects
}
\author{
Mikko Vihtakari
}
