% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zooplankton_data.R
\name{read_zooplankton_data}
\alias{read_zooplankton_data}
\title{Read NPI's zooplanton data from an Excel sheet}
\usage{
read_zooplankton_data(data_file, sheet = 1, dataStart = 11, dataEnd = 266,
  dataCols = NULL, control_stations = TRUE, output_format = "as.Date",
  add_coordinates = FALSE, species_info_cols = c("group", "species",
  "stage", "size_op", "length"), remove_missing = TRUE)
}
\arguments{
\item{data_file}{Path to the Excel file containing zooplankton data}

\item{sheet}{The name or index of the sheet to read the zooplanton data from. See \code{\link[openxlsx]{read.xlsx}}}

\item{dataStart}{The row number where zooplanton data starts from}

\item{dataEnd}{The row number where zooplankton data ends}

\item{dataCols}{Optional numeric index indicating the column numbers that contain zooplankton data}

\item{control_stations}{Logical indicating whether station names should be controlled against a list of standardized station names (see \code{\link{STATIONS}}).}

\item{output_format}{Output formar for date. See \code{\link{convert_dates}}.}

\item{add_coordinates}{If \code{TRUE} coordinates will be added to metadata from the list of standardized station names.}

\item{species_info_cols}{Character vector specifying the species information columns that should be preserved.}

\item{remove_missing}{Logical indicating whether species with column sums of 0 should be removed from the output.}
}
\value{
Returns a list of class \code{ZooplanktonData}. The list contains 3 data frames: \code{$data} (abundance data), \code{$meta} (meta-data), and \code{$splist} (species information).
}
\description{
Reads NPI's standard format zooplankton data
}
\author{
Mikko Vihtakari
}
