% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.ZooplanktonData.R
\name{subset.ZooplanktonData}
\alias{subset.ZooplanktonData}
\title{Subset a ZooplanktonData object}
\usage{
\method{subset}{ZooplanktonData}(x, type, ..., remove_zeros = FALSE,
  drop_levels = TRUE)
}
\arguments{
\item{x}{a \link[=read_zooplankton_data]{ZooplanktonData} object.}

\item{type}{Character defining whether meta-data ("meta") or species list ("splist") should be used in subsetting.}

\item{...}{Arguments passed to the \code{\link[base]{subset}} function.}

\item{remove_zeros}{Logical indicating whether columns/species (if \code{type = "meta"}) or rows/samples (if \code{type = "splist"}) with zero sums should be removed from the dataset.}

\item{drop_levels}{Logical indicating whether unused levels should be dropped from \code{$meta} and \code{$splist} data frames. See \code{\link[base]{droplevels}}.}
}
\value{
Returns a subsetted \link[=read_zooplankton_data]{ZooplanktonData} object.
}
\description{
Generic subset method for ZooplanktonData objects. Returns a subset of a Zooplankton object based on meta-data or species information defined by the \code{type} argument.
}
\details{
See the the \code{\link[base]{subset}} function for details.
}
\seealso{
Other ZooplanktonData: \code{\link{merge_zooplankton_data}},
  \code{\link{print.ZooplanktonData}},
  \code{\link{read_zooplankton_data}},
  \code{\link{summarize_zooplankton_data}}
}
\author{
Mikko Vihtakari
}
\concept{ZooplanktonData}
