% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_zooplankton_data.R
\name{summarize_zooplankton_data}
\alias{summarize_zooplankton_data}
\title{Summarize ZooplanktonData species abundances based on information from meta-data or species list}
\usage{
summarize_zooplankton_data(obj, sp_group = NULL, meta_group = NULL,
  meta_group_method = "depth_mean", biomass = FALSE,
  remove_noncont = FALSE, warnings = TRUE)
}
\arguments{
\item{obj}{\link[=read_zooplankton_data]{ZooplanktonData} object}

\item{sp_group}{character indicating the new names of species matrix names. \strong{Sums up} species names based on the new name. Can be used to collate species to remove stages ("species") or to collate species based on origin ("origin"; Atlantic vs Arctic). Use \code{NULL} if you do not want to change species matrix names.}

\item{meta_group}{character vector of column names in \code{$meta}, which should be used to summarize the species data. Typically use c("expedition", "station", "lon", "lat", "lon.utm", "lat.utm", "date") for multinet data to summarize over the entire water column. Use \code{NULL} if you do not want to summarize species data based on meta data.}

\item{meta_group_method}{character giving the method which should be used to summarize species data. Options: 
\itemize{
  \item \strong{"mean"} takes a simple mean of based on values listed in \code{meta_group} argument.
  \item \strong{"depth_mean"} takes a depth averaged abundance/biomass using the values listed in \code{meta_group} argument. Requires \code{from} and \code{to} columns in \code{$meta}. The values for each species are multiplied by depth interval, summed up by \code{meta_group} and divided by summed depth interval. The unit is the original unit (n/m3 or weight/m3). 
  \item \strong{"depth_sum"} same than above, but values are not divided by summed depth interval. Represents depth integrated" values. The unit is n/m2 if the original unit was n/m3.
  \item \strong{"sum"} sums up ("integrates") abundances/biomasses for each species by \code{meta_group}
  \item \strong{"total_sum"} ignores species and sums up all abundance/biomass by \code{meta_group}. 
}}

\item{biomass}{logical indicating whether abundances should be converted to biomass. Requires the \code{biomass_conv} column in \code{obj$splist}.}

\item{remove_noncont}{logical indicating whether non-continuous casts should be removed if \code{meta_group_method} = "depth_mean".}

\item{warnings}{logical indicating whether the function should print warnings if it detects problems with data}
}
\value{
Returns a list of class \code{ZooplanktonData}. The list contains 3 data frames: \code{$data} (abundance data), \code{$meta} (meta-data), and \code{$splist} (species information).
}
\description{
Averages \link[=read_zooplankton_data]{ZooplanktonData} objects by grouping varaibles from meta-data (\code{$meta}) or species list (\code{$splist}). Can be used to convert abundances to biomasses, to sum up species abundances to remove stages and to summarize species data over the entire water column or regions.
}
