% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{ts_plot}
\alias{ts_plot}
\title{Temperature-salinity (T-S) plot from a data frame using ggplot2}
\usage{
ts_plot(dt, temp_col = "theta", sal_col = "salinity", xlim = NULL,
  ylim = NULL, color = "watertype", zoom = TRUE, margin_distr = FALSE,
  nlevels = 6)
}
\arguments{
\item{dt}{data frame.}

\item{temp_col}{character specifying the column name that contains temperature information. Make sure that the temperatures are potential temperatures (theta) instead of in-situ tempretatures.}

\item{sal_col}{character specifying the column name that contains salinity information. Make sure that the salinities are practical salinities (theta).}

\item{color}{character indicating the mapping variable for color. If \code{"watertype"} the cots are colored based on water type using a prededined palette.}

\item{zoom}{logical indicating whether the x- and y-axis should be limited to data. If \code{FALSE}, the entire water mass diagram from Cottier et al. is shown.}

\item{margin_distr}{logical indicating whether kernel density estimates of temperature and salinity should be added to margins of the plot.}

\item{nlevels}{Number of automatically-selected isopycnal levels. Set to 0 to remove isopycnals.}
}
\description{
Makes a T-S plot from salinity and temperature data. Optimized for Kongsfjorden.
}
\references{
Cottier, F., Tverberg, V., Inall, M., Svendsen, H., Nilsen, F., Griffiths, C., 2005. Water mass modification in an Arctic fjord through cross-shelf exchange: The seasonal hydrography of Kongsfjorden, Svalbard. J. Geophys. Res. 110, C12005. doi:10.1029/2004JC002757
}
\seealso{
\code{\link{define_water_type}}
}
