% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{warn_vec}
\alias{warn_vec}
\title{Return character vector elements or length depending on length}
\usage{
warn_vec(x, len = 5)
}
\arguments{
\item{x}{character vector}

\item{len}{integer specifying the cut value. If length of \code{x} <= \code{len}, a character vector will be returned. Otherwise a length as character.}
}
\description{
This internal function returns character vector elements or length as a character depending on the length of \code{x}. Useful in \link[base]{stop} and \link[base]{warning} functions.
}
