% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coln_search_words.R
\name{coln_search_words}
\alias{coln_search_words}
\title{A list of search words used to find column names in Excel sheets}
\usage{
coln_search_words(column)
}
\arguments{
\item{column}{A required column name as character string}
}
\description{
A switch list containing search words used in \code{guess_colname} function
}
\details{
The function accepts following required column names: \code{expedition}, \code{station}, \code{type}, \code{sample_name}, \code{longitude}, \code{latitude}, \code{date}, \code{bottom_depth}, \code{gear}, \code{from}, \code{to}, \code{filtered_volume}, \code{responsible} and \code{comment}.
}
\author{
Mikko Vihtakari, Conrad Helgeland
}
