% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_species_data.R
\encoding{UTF-8}
\name{compile_species_data}
\alias{compile_species_data}
\title{Compile species data from ice cores}
\usage{
compile_species_data(dat, splist, core_cols = c("expedition", "station",
  "core.id"), sp_col = "species", ab_col = "abundance", unit_col = "unit",
  start_col = "from", end_col = "to", ice_thick_col = "ice",
  add_cols = c("gear", "longitude", "latitude", "date", "bottom_depth",
  "ice.type", "snow"), convert_unit = list(from = "1/L", to = "1/m2"),
  summarise_cores = TRUE, return_summary = TRUE, round_digits = 0)
}
\arguments{
\item{dat}{data frame containing ice core species data to be combined. Must be in long format.}

\item{splist}{species list. Preferably from \code{\link{confirm_species_list}}.}

\item{core_cols}{A character vector specifying columns that separate individual ice-cores.}

\item{sp_col}{Character specifying the name of the species column in \code{dat}.}

\item{ab_col}{Character specifying the name of the abundance column in \code{dat}.}

\item{unit_col}{Character specifying the name of the unit for the abundance column in \code{dat}.}

\item{start_col}{Character argument specifying the name of the column in \code{dat} for the section start for the ice core in centimeters or meters.}

\item{end_col}{Character argument specifying the name of the column in \code{dat} for the section end for the ice core in centimeters or meters.}

\item{ice_thick_col}{Character argument specifying the name of the column in \code{dat} for ice thickness preferably in \strong{centimeters}}

\item{add_cols}{A character vector containing names of the additional columns that should be included in the output. These columns are not used by the function otherwise. Additional columns cannot contain several unique values due to summarizing process.}

\item{convert_unit}{List or \code{NULL}. Should the function convert supplied abundances using the \code{\link{convert_abundace}} function, a list giving \code{from} and \code{to} units have to be supplied. If conversion should not be made, use \code{NULL} (default). See details for alternatives.}

\item{summarise_cores}{Logical. Should core sections be summarized to a continuous core? If \code{TRUE} (default), all sections are summarized, \code{start_col} and \code{end_col} removed from the output, and a \code{$core_type} column added. If \code{FALSE} core sections are returned as they are, but duplicate species names are summed up.}

\item{return_summary}{Logical. Should summary statistics be returned?}

\item{round_digits}{Number of digits \code{start_col}, \code{end_col} and \code{ice_thick_col} values should be rounded to. Set to \code{NULL} to avoid rounding. See \code{\link[base]{round}} for details.}
}
\value{
Returns a data frame with columns specified in arguments. Drops any unspecified column.
}
\description{
Compiles species data across a larger dataset with duplicate species names as defined by a species list from \code{\link{confirm_species_list}} function.
}
\details{
Built for ice-algae taxonomy data, but could be modified to work with any taxonomy data including plankton nets. The function is currently broken for wider use.

Implemented \strong{convert_unit} argument alternatives:
\itemize{
  \item \strong{\code{NULL}}: no conversion. Adundances summed up as they are (see \code{summarise_cores} argument).
  \item \strong{\code{list(from = "per", to = "per")}}: percentage abundances are summed up according to the \code{summarise_cores} argument and scaled to 100 \% (sum for an ice core or a section adds up to 100 \%).
  \item \strong{\code{list(from = "rel", to = "rel")}}: relative abundances are summed up by taking a mean for duplicate species or entire ice cores (if \code{summarise_cores = TRUE}).
  \item \strong{\code{list(from = "rel", to = "per")}}: relative abundances are first summed up and a percentage of their contribution is calculated. Abundances add up to 100 \% for an ice core or a section depending on the \code{summarise_cores} argument.
  \item \strong{\code{list(from = "1/L", to = "1/m2")}} (or any other variations passed to \code{\link{convert_abundace}}): Individuals per litre values converted to individuals per square metre before summing up abundances (see \code{summarise_cores} argument)
  }
}
\author{
Mikko Vihtakari
}
