% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_metadata.R
\name{export_metadata}
\alias{export_metadata}
\title{Export meta-data to NPI database input format}
\usage{
export_metadata(meta_file, sheet = 1, expedition = "Expedition",
  station = "Station", type = "Sample.type", sample_name = "Sample.name",
  longitude = "Longitude.(decimals)", latitude = "Latitude.(decimals)",
  date = "Sampling.date.(UTC)", bottom_depth = "Bottom.depth.(m)",
  gear = "Gear", from = "Sampling.depth.(m).from",
  to = "Sampling.depth.(m).to", filtered_volume = "Filtered.volume",
  responsible = "Responsible.person", comment = "Comment",
  additional = NULL)
}
\description{
Exports meta-data from Excel files to the format accepted by NPI data managers
}
