% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_metadata.R
\name{export_metadata}
\alias{export_metadata}
\title{Export meta-data to NPI database input format}
\usage{
export_metadata(meta_file, sheet = 1, expedition = "Expedition",
  station = "Station", type = "Sample.type", sample_name = "Sample.name",
  longitude = "Longitude.(decimals)", latitude = "Latitude.(decimals)",
  date = "Sampling.date.(UTC)", bottom_depth = "Bottom.depth.(m)",
  gear = "Gear", from = "Sampling.depth.(m).from",
  to = "Sampling.depth.(m).to", filtered_volume = "Filtered.volume",
  responsible = "Responsible.person", comment = "Comment",
  additional = NULL, guess_colnames = FALSE)
}
\arguments{
\item{meta_file}{File name of the meta-data table. Currently only Excel files are supported.}

\item{sheet}{sheet number or name where meta-data are located. See \code{\link[openxlsx]{read.xlsx}}}

\item{expedition}{Column name specifying the name of the expedition.}

\item{station}{Column name specifying the station name.}

\item{type}{Column name specifying the sample type column.}

\item{sample_name}{Column name specifying the sample name column.}

\item{longitude}{Column name specifying the longitude column. Coordinates should be given in decimal degrees.}

\item{latitude}{Column name specifying the latitude column. Coordinates should be given in decimal degrees.}

\item{date}{Column name specifying the sampling time. The information will be transformed to ISO 8601 standard format.}

\item{bottom_depth}{Column name specifying the bottom depth during sampling.}

\item{gear}{Name of the column specifying the sampling gear.}

\item{from}{Name of the column specifying the depth from which sampling was started.}

\item{to}{Name of the column specifying the depth from which sampling was ended.}

\item{filtered_volume}{Name of the column specifying filtered volume in ml for filtered samples.}

\item{responsible}{Name of the column specifying the responsible persons for the sampling.}

\item{comment}{Name of the column specifying comments. Any non-numeric values in \code{from} and \code{to} will be transferred to this column.}

\item{additional}{Additional columns to be included in meta-data. Must be specified as a character vector, which lists exact column names to be included.}

\item{guess_colnames}{Logical indicating whether fuzzy matching (\code{\link[base]{agrep}}) should be used to guess column names.}
}
\value{
Returns a list of class \code{MetaData} that contains modified meta-data records (\code{$meta}), removed meta-data records (\code{$deleted}), file name to be used when saved for database import (\code{file_id}) and a data frame containing row numbers of duplicates in the original Excel sheet (\code{$duplicates}).
}
\description{
Exports meta-data from Excel files to the format accepted by NPI data managers
}
\details{
All column names should be specified as character strings of length 1. The column names refer to the meta-data table (\code{dt}).

Columns \code{expedition}, \code{station}, \code{type}, \code{sample_name}, \code{longitude}, \code{latitude}, \code{date}, \code{bottom_depth}, \code{gear}, \code{from}, \code{to}, \code{filtered_volume}, \code{responsible} and \code{comment} are required. Each of these column arguments have a reasonble "guess" column name (see \strong{Usage}) that should match the column names when read from an Excel file. You can use \code{guess_colnames = TRUE} to make the function guess column names, if they differ somewhat from the original column names listed in \strong{Usage}.
}
\examples{
\donttest{
## Read meta-data and let the function to find errors in it:
x <- export_metadata("Samplelog MOSJ 2015.xlsx")

## Meta-data reading follows openxlsx syntax:
x <- export_metadata("GlacierFront_2017_Samplelog_20171024.xlsx", sheet = "SAMPLELOG",
guess_colnames = TRUE)
}
}
