% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_metadata.R
\name{guess_colname}
\alias{guess_colname}
\title{param cols columns to be guessed
param df data frame with column names}
\usage{
guess_colname(cols = required_cols, df = dt)
}
\description{
param cols columns to be guessed
param df data frame with column names
}
