% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_species_list.R
\name{make_species_list}
\alias{make_species_list}
\title{Generate a species list containing unique levels from a taxonomy dataset}
\usage{
make_species_list(
  dat,
  sp.col = "species",
  ab.col = NULL,
  st.col = NULL,
  exp.col = NULL
)
}
\arguments{
\item{dat}{Data frame containing taxonomy information. Must be in long format.}

\item{sp.col}{Character specifying the column name that contains species names in \code{dat}.}

\item{ab.col}{Character specifying the column name that contains species abundances in \code{dat}.}

\item{st.col}{Character specifying the column name that contains station names in \code{dat}.}

\item{exp.col}{Character specifying the column name that contains expedition in \code{dat}.}
}
\value{
Returns \code{speciesList} object containing a list of unique species entries in \code{dat} (\code{$data}) and summary information (\code{$sum_info}). \code{$data} contains following columns:
\itemize{
 \item \strong{\code{or_name}} The original species names (as unique levels in \code{dat}).
 \item \strong{\code{expedition}} Expeditions during which the species occurred. Several expeditions are separated by \code{;}.
 \item \strong{\code{ab}} Summed abundance for a species entry.
 \item \strong{\code{n}} Number of stations where the species name was encountered (n)
 \item \strong{\code{perAb}} Summed abundance for a species entry as a percentage of total summed abundance (\code{$sum_info$abSum}).
 \item \strong{\code{fo}} Frequency of occurrence for a species entry (total number of stations is given in \code{$sum_info$N}).
 \item \strong{\code{search_term}} A tidied search term for \code{\link{confirm_species_list}} function.
 \item \strong{\code{type}} Type of species (spores, symbionts and epiphytes are separated, otherwise \code{NA}).
 \item \strong{\code{certainty}} Certainty of species identification (includes sp., spp., aff., cf. and unknown).
 }
}
\description{
Generate a data frame containing unique species names, summed abundances and search terms to be used by \code{\link{confirm_species_list}}.
}
\seealso{
\code{\link{confirm_species_list}} for confirming the species names against a database.
}
\author{
Mikko Vihtakari
}
