% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zooplankton_data.R
\name{read_zooplankton_data}
\alias{read_zooplankton_data}
\title{Read NPI's zooplanton data from an Excel sheet}
\usage{
read_zooplankton_data(data_file, sheet = 1, dataStart = 11,
  dataEnd = 1000, dataCols = NULL, control_stations = TRUE,
  output_format = "as.Date", add_coordinates = FALSE,
  control_species = list(species = "species", stage = "stage", length =
  "length"), species_info_cols = c("species", "stage", "length"),
  lookup_cols = NULL, remove_missing = TRUE)
}
\arguments{
\item{data_file}{Path to the Excel file containing zooplankton data}

\item{sheet}{The name or index of the sheet to read the zooplanton data from. See \code{\link[openxlsx]{read.xlsx}}}

\item{dataStart}{The row number where zooplanton data starts from}

\item{dataEnd}{The row number where zooplankton data ends}

\item{dataCols}{Optional numeric index indicating the column numbers that contain zooplankton data}

\item{control_stations}{Logical indicating whether station names should be controlled against a list of standardized station names (see \code{\link{STATIONS}}).}

\item{output_format}{Output formar for date. See \code{\link{convert_dates}}.}

\item{add_coordinates}{If \code{TRUE} coordinates will be added to metadata from the list of standardized station names.}

\item{control_species}{Either \code{NULL} or a named list giving the name of species, stage and length columns in data, which will be used to control species names against a list of standardized zooplanton species (see \code{ZOOPL}). Recommended. Seetting this to \code{NULL} may lead to unexpected behaviour.}

\item{species_info_cols}{Character vector specifying the names of species information columns that should be preserved. If \code{control_species = TRUE}, the infomation will be used to match species entries agains the zooplankton lookup list (see \code{ZOOPL}).}

\item{lookup_cols}{Character vector specifying the names of columns from the zooplankton lookup list (\code{ZOOPL}) that should be returned together with \code{species_info_cols}. If \code{NULL} (default), only \code{species_info_cols} will be returned. Has no effect, if \code{control_species = FALSE}.}

\item{remove_missing}{Logical indicating whether species with column sums of 0 should be removed from the output.}
}
\value{
Returns a list of class \code{ZooplanktonData}. The list contains 3 data frames: \code{$data} (abundance data), \code{$meta} (meta-data), and \code{$splist} (species information).
}
\description{
Reads NPI's standard format zooplankton data
}
\details{
The function sums up duplicate species entries for each sample. 

The function is currently relatively unstable and most likely requires manual debugging for each dataset.
}
\author{
Mikko Vihtakari
}
